Edit the Navbar
I edited the brand and the links of the Navbar. The links of 'All',
'Other' and different rooms will link to the FurnitureList component,
which will be designed to show different set of products based on the
catogories, and the wishlist is equivalent to the favorites page at this
moment.